// src/utils/enums/MessageEnum.ts
enum UserEnum {
  USER_NOT_FOUND = 'User not found',
  WRONG_PASSWORD = 'Wrong Password',
  LOGIN_SUCCESSFULLY = 'Logged in Successfully',
  USER_SAVE_SUCCESS = 'User Saved Successfully',
  INVALID_TOKEN = 'Token is invalid or expired',
  USER_INFO = 'User information',
  INVALID_LOGIN = 'Login is invalid',
}
export enum ProductEnum {
  PRODUCT_CREATE_FAILED = 'Product create failed',
  WRONG_PASSWORD = 'Wrong Password',
  LOGIN_SUCCESSFULLY = 'Logged in Successfully',
  PRODUCT_SAVE_SUCCESS = 'Product Saved Successfully',
  INVALID_TOKEN = 'Token is invalid or expired',
  USER_INFO = 'User information',
  INVALID_LOGIN = 'Login is invalid',
  PLAN_CREATE_FAILED = 'Plan failed to create',
  PLAN_CREATED_SUCCESSFULLY = 'Plan created successfully',
  CUSTOMER_NOT_CREATED = 'Customer not created',
  CUSTOMER_CREATED_SUCCESSFULLY = 'Customer created successfully',
  PAYMENT_METHOD_NOT_CREATED = 'Payment method not created',
  PAYMENT_METHOD_CREATED_SUCCESSFULLY = 'Payment method created successfully',
}
export default UserEnum;